event-catalog-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.2
pandas==2.2.2
